+++ apps/docs/src/app/core/component-docs/checkbox/examples/checkbox-default-example.component.ts
@@ -3,7 +3,8 @@ import { Component } from '@angular/core';
 @Component({
     selector: 'fd-checkbox-default-example',
     template: `
-        <fd-checkbox label="Option 1" [(ngModel)]="checkboxValue" required="true"></fd-checkbox>
+        <label fd-form-label for="defaultCheckbox" [required]="true">Default checkbox</label>
+        <fd-checkbox label="Option 1" [(ngModel)]="checkboxValue" required="true" inputId="defaultCheckbox"></fd-checkbox>
         Value: {{ checkboxValue }}
     `
 })